telephan_dashboard/requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
bcrypt==4.2.0
python-dotenv==1.0.1
